Account for hairline mode in rect op bounds for clipping
Disables the preApply() clip check for rects that are subpixel since the
bounds will not be correct, and avoids pre-applying clips to rects if
the result would turn into a subpixel draw.

Also disables hairline mode for rects that are not fully anti-aliased,
which showed up in the fast_constraint GMs where we'd tile an image and
the bottom/right edges were actually subpixel. Since the objective of
per-edge AA is for tiling, we don't want those subpixel edges to be
outset more than the original geometry would have been in a non-tiling
case.

Bug: skia:12164, chromium:1210170
Change-Id: I3bca4fc3e63b6f86f1da1ea7a5e5873d221755f7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/426437
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>