Never disable multisample with DMSAA
DMSAA used to disable multisample in order to use coverage ops, but it
turns out that disabling multisample while rendering to an MSAA target
is very expensive. This CL updates DMSAA to always prefer the MSAA op
unless a coverage version exists that also knows how to render when
multisampling is enabled. We will follow it up with updates to
coverage ops that make them MSAA compatible.

Bug: skia:11396
Change-Id: I8e4ea763710d96593e08877b100eb7777268d7eb
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/427657
Commit-Queue: Chris Dalton <csmartdalton@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>