Bail on AAStrokeRectOp of there are unequal, subpixel widths
Unequal, subpixel stroke widths had a known rendering issue that we
just accepted. But with DMSAA compatibility coming up, these will
potentially pop visually when turning MSAA on and off. Fall back on
path rendering in this case.

A side effect of this change is that the insets and outsets are now
always equal in X and Y, so we also update the vertex generation code
to not distinguish between X and Y.

Bug: skia:11396
Change-Id: Iaf0ce73c228ad67e55aa5c004dbdf18bb2134594