Fix AAStrokeRectOp viewport clipping for nested rects
The code analyzes the nested rects in order to surmise stroke sizes.
When we clipped the nested rects before deducing a stroke size, it
led to a stroke size of 0 which made coverage 0 all throughout.

This CL changes the code to figure a stroke size before clipping, and
to outset the clip for the outer rect by the stroke size.

TBR=bsalomon@google.com

Bug: skia:12211
Change-Id: I15dde092df9b0cbd56c6023b91cd2d3430fde10a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/428656
Reviewed-by: Chris Dalton <csmartdalton@google.com>
Commit-Queue: Chris Dalton <csmartdalton@google.com>