Fix AAStrokeRectOp viewport clipping for nested rects
The code looks at the nested rects in order to deduce the stroke
sizes. When we clipped the nested rects before dedcuing a stroke
width, it led to a stroke width of 0 and a uniform coverage of 0.

This CL changes the code to figure a stroke size before clipping, and
to outset the clip for the outer rect by the stroke size.

Bug: skia:12211
Change-Id: I15dde092df9b0cbd56c6023b91cd2d3430fde10a