Turn off caps.multisampleDisableSupport()
Disabling multisample while rendering to an MSAA target is a dwindling
feature that comes with a heavy performance cost on some devices, and
many devices don't support it at all. This CL just flips a bit so we
never disable multisample. After a few days or weeks of soak time, we
will come back and yank out all the code that supports disabling
multisample.

Bug: skia:12196
Change-Id: Iae54d25d0e28e24eda6295d7155e5c9fbb398081
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/429299
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Chris Dalton <csmartdalton@google.com>