Take the dest color as an input to SkBlenderBase::asFragmentProcessor.
Previously, the dest color was implicitly the surface that we were
drawing onto. When setting the blender on an SkPaint, this is still the
case. For other uses, e.g. a SkBlendImageFilter, the caller may want to
supply an arbitrary input FP.

Bug: skia:12085, skia:12205
Change-Id: I254be006a0f374711fa64f49e0ba339578d8d83a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/430059
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>