Remove the viewMatrix uniform from tessellation's BoundingBoxShader
This can be a instance attrib instead, which allows better batching
now that we transform the stencil geometry on CPU. We still keep color
as a uniform for this CL, since that can be faster on some devices
than varying color. In the future this can be configurable if we need
different colors in order to batch.

Bug: skia:12258
Change-Id: Ib8a4384f44586ffc80e72ca1cea8e9e49c778cdf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/433556
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Chris Dalton <csmartdalton@google.com>