SkScalerContext handle stroke changing path effect
A path effect is allowed to modify the stroke, which may affect whether
or not the final path is hairline or not. Handle this similarly to
SkPaint::getFillPath by taking an out parameter which states if the path
is hairline (should be zero stroked) or is an actual fill path.

This simplifies the internal handling of hairlines in SkScalerContext as
well as making the result more in line with the way hairlines are
handled elsewhere. It should also make the removal of
SkPaint::kStrokeAndFill_Style simpler.

Change-Id: I8ae8e3d6ee8a9f686f983cbcf723925b11c4a707
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/437020
Reviewed-by: Mike Reed <reed@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>